pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-integration-tests"
version = "0.2.0"
description = "Cross-repo integration test suite for AumOS Enterprise platform"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "aumos-common>=0.1.0",
    "aumos-proto>=0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "testcontainers[postgres,kafka,redis]>=4.7.0",
    "docker>=7.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "confluent-kafka>=2.3.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
integration = [
    "testcontainers[postgres,kafka,redis]>=4.7.0",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "pact-python>=2.2.0",
    "pytest-benchmark>=4.0.0",
    "confluent-kafka>=2.3.0",
]
contracts = [
    "pact-python>=2.2.0",
    "httpx>=0.27.0",
]
performance = [
    "pytest-benchmark>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "phase0: Foundation integration tests",
    "phase1: Data Factory integration tests",
    "phase2: DevOps + Trust integration tests",
    "smoke: Quick smoke tests",
    "integration: Requires real infrastructure containers (Testcontainers)",
    "contract: Pact consumer contract tests",
    "chaos: Chaos and fault-injection tests",
    "performance: Performance baseline tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
